ref 汎用関数
